Add files via upload
The original purpose of this script is to create read a csv file containing user info and start date, depending on the startdate of the user account, a scheduled task is created or the script is executed. 

The script is created for a hybrid environment, the user account is created on-prem, then synced to Entra and another access is granted to DevOps project. The script also enables remote mailbox for the user account